refactor(vscode): untangle webview host and extension entrypoints
Extract shared SSE, active-editor, message fan-out, and retry helpers so
Chat/Agent Manager/Session Editor no longer duplicate host plumbing.
Split the webview god file into route, fetch, loading, command, and
notification modules, and thin extension.ts by moving sidebar placement
and OpenCode status reporting into focused modules. Add focused unit
tests and clearer contributor onboarding in the package README/docs.
